🎨 [REFACTORING] CounterState 폴더 위치 변경 및 import 정리 #243
