+++ src/store/cartModalState.js
@@ -1,9 +1,11 @@
 import { atom /* selector */, selector } from 'recoil';
 import { recoilPersist } from 'recoil-persist';
-import { productListFamily } from './productListState';
-import { countState } from './CounterState';
 
-// 영속성 
+import { productListFamily } from './productListState';
+
+import { countState } from '@/components/Counter/@recoil/counterState.js';
+
+// 영속성
 const { persistAtom } = recoilPersist({
   key: 'recoil-persist', // this key is using to store data in local storage
   storage: localStorage, // configurate which stroage will be used to store the data
